Implemented ship movement using WASD, fixed some bugs
git-svn-id: https://svn.fe.up.pt/repos/lcom1920-t6g02@157 812d746c-baee-4303-96b4-7e37b0e189f4

+++ proj/keyboard.c
@@ -6,7 +6,7 @@
 #include "kbc.h"
 
 static int kbd_hook_id = KBD_IRQ;
-extern int scancode;
+extern uint8_t scancode;
 
 int kbd_subscribe_int(uint8_t *bit_no) {
   if (bit_no == NULL) {
